+++ CHANGELOG.md
@@ -9,6 +9,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - `it-ticket-v1`: implemented addon for AdE e-receipt format 
+- `bill`: line discount and charge `base` property, to use instead of the line sum in order to comply with EN16931.
+- `bill`: line Charge support for Quantity and Rate special cases for charges like tariffs that result in a fixed amount base on a rate, like, 1 cent for every 100g of sugar.
+
+### Changed
+
+- `bill`: line totals will be rounded to currency precision for presentation only
+- `bill`: document Discount and Charge base and amounts always rounded to currency's precision
+- `bill`: line Discount and Charge base and amounts always rounded to currency's precision
+- `tax`: renamed rounding rules `sum-then-round` to `precise`, and `round-then-sum` to `currency`, to more accurately reflect their objectives.
+- `bill`: `currency` rounding rule implies that line totals will be calculated with the currency's precisions, bringing closer alliance with EN16931 requirements.
 
 ## [v0.211.0] - 2025-02-28